streamlit-prototype/requirements.txt
agno>=1.0.0
streamlit
python-dotenv
cerebras-cloud-sdk
